cart delete functioning

+++ app/api/cart_routes.py
@@ -25,7 +25,7 @@ def index():
 #/api/cartproduct/id
 #/api/cart/product/productid
 # delete from cart
-@cart_routes.route('/product/<int:id>', methods=['DELETE'])
+@cart_routes.route('/cartProduct/<int:id>', methods=["DELETE"])
 @login_required
 def delete_from_cart(id):
     # userId = current_user.id
@@ -35,9 +35,16 @@ def delete_from_cart(id):
     # matching cart id and product id
 
     print('checking backend', id)
-    product_to_delete = CartProduct.query.filter(id == CartProduct.product_id).one()
-    print('product to delete---------', product_to_delete)
-    db.session.delete(product_to_delete)
-    db.session.commit()
+    # product_to_delete = CartProduct.query.filter(id == CartProduct.product_id).one()
 
-    return redirect('/api/cart')
+    #testing for delete
+    product_to_delete = CartProduct.query.get(id)
+
+    print('product to delete---------', product_to_delete)
+    print('product to delete id', product_to_delete.id)
+    if product_to_delete is not None:
+        db.session.delete(product_to_delete)
+        db.session.commit()
+        return {'message': 'Successfully Deleted'}
+
+    return 'Cart Product not found'
